EE Staging: - Fixed: #21905: Time of date of Schedule Marge looks strange - Fixed: #21943: The backup wasn't been created during work Schedule Merge - Fixed: Fatal when trying delete backup that doesn't exist - Code refactoring

+++ app/code/core/Enterprise/Staging/controllers/Adminhtml/Staging/BackupController.php
@@ -156,11 +156,13 @@ class Enterprise_Staging_Adminhtml_Staging_BackupController extends Enterprise_S
         $backup         = $this->_initBackup();
         $redirectBack   = false;
 
-        try{
-            $backup->setIsDeleteTables(true);
-            $backup->delete();
-        } catch (Exception $e) {
-            $redirectBack = true;
+        if ($backup) {
+            try{
+                $backup->setIsDeleteTables(true);
+                $backup->delete();
+            } catch (Exception $e) {
+                $redirectBack = true;
+            }
         }
 
         if ($redirectBack) {
